+++ src/com/magento/idea/magento2plugin/util/magento/plugin/GetTargetClassNamesByPluginClassName.java
@@ -2,47 +2,53 @@
  * Copyright © Magento, Inc. All rights reserved.
  * See COPYING.txt for license details.
  */
+
 package com.magento.idea.magento2plugin.util.magento.plugin;
 
 import com.intellij.openapi.project.Project;
 import com.intellij.psi.search.GlobalSearchScope;
 import com.intellij.util.indexing.FileBasedIndex;
 import com.magento.idea.magento2plugin.stubs.indexes.PluginIndex;
-
+import com.magento.idea.magento2plugin.stubs.indexes.data.PluginData;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.List;
 import java.util.Set;
 
 /**
- * Returns all targets class names for the plugin
+ * Returns all targets class names for the plugin.
  */
 public class GetTargetClassNamesByPluginClassName {
-    private static GetTargetClassNamesByPluginClassName INSTANCE = null;
-    private Project project;
+    private final Project project;
 
-    public static GetTargetClassNamesByPluginClassName getInstance(Project project) {
-        if (null == INSTANCE) {
-            INSTANCE = new GetTargetClassNamesByPluginClassName();
-        }
-        INSTANCE.project = project;
-        return INSTANCE;
+    public GetTargetClassNamesByPluginClassName(final Project project) {
+        this.project = project;
     }
 
-    public ArrayList<String> execute(String currentClassName) {
-        ArrayList<String> targetClassNames = new ArrayList<>();
-        Collection<String> allKeys = FileBasedIndex.getInstance()
+    /**
+     * Get current class names for the plugin.
+     *
+     * @param currentClassName plugin class name.
+     * @return list of plugin classes
+     */
+    public List<String> execute(final String currentClassName) {
+        final List<String> targetClassNames = new ArrayList<>();
+        final Collection<String> allKeys = FileBasedIndex.getInstance()
                 .getAllKeys(PluginIndex.KEY, project);
 
-        for (String targetClassName : allKeys) {
-            List<Set<String>> pluginsList = FileBasedIndex.getInstance()
-                    .getValues(com.magento.idea.magento2plugin.stubs.indexes.PluginIndex.KEY, targetClassName, GlobalSearchScope.allScope(project));
+        for (final String targetClassName : allKeys) {
+            final List<Set<PluginData>> pluginsList = FileBasedIndex.getInstance()
+                    .getValues(
+                        PluginIndex.KEY,
+                        targetClassName,
+                        GlobalSearchScope.allScope(project)
+                    );
             if (pluginsList.isEmpty()) {
                 continue;
             }
-            for (Set<String> plugins : pluginsList) {
-                for (String plugin : plugins) {
-                    if (!plugin.equals(currentClassName)) {
+            for (final Set<PluginData> plugins : pluginsList) {
+                for (final PluginData plugin : plugins) {
+                    if (!plugin.getType().equals(currentClassName)) {
                         continue;
                     }
                     targetClassNames.add(targetClassName);